Rename `lib/bruter.h` to `lib/utils`
I did this because the filename bruter is not suited well to
what we are doing. The file contains utils or functions that
do misc stuff like checking wheter a user has read permission
and it doesn't actually brute force directories.

The next commit will include a new file `lib/bruter.h` that
will the brute force thingie.