Don't store file contents in an array.
Instead of storing file contents in an array, we use getline()
then we format the host and we feed to request, by adding to
it the work queue. So no need to actually store this value.

This approach however, produces a strage issue, the program
reports 200 OK on links that are 404 Not Found. I still yet
to figure out what is going on with all of that.

Furthermore, I have to sleep() for a while before calling
thpool_add_work(). Which I have to figure out why since it's
the only way I got the program to acutally run properly.